Clean up existing retriever
