+++ tsdat/io/retrievers.py
@@ -6,20 +6,17 @@ import xarray as xr
 from pydantic import BaseModel, Extra
 from typing import Any, Dict, List, Pattern, cast
 from ..config.dataset import DatasetConfig
-from .base import Retriever, DataReader, DataConverter
+from .base import Retriever, DataReader, RetrievedVariable
 
 __all__ = ["DefaultRetriever"]
 
 logger = logging.getLogger(__name__)
 
 
-class RetrievedVariable(BaseModel, extra=Extra.forbid):
-    name: str
-    data_converters: List[DataConverter] = []
-
-
 class InputKeyRetrieverConfig:
-    def __init__(self, input_key: str, retriever: "DefaultRetriever") -> None:
+    """Gathers variable retrieval rules for the given input key."""
+
+    def __init__(self, input_key: str, retriever: Retriever) -> None:
         self.coords: Dict[str, RetrievedVariable] = {}
         self.data_vars: Dict[str, RetrievedVariable] = {}
 
@@ -73,16 +70,6 @@ class DefaultRetriever(Retriever):
     """A dictionary of DataReaders that should be used to read data provided an input
     key."""
 
-    coords: Dict[str, Dict[Pattern, RetrievedVariable]]  # type: ignore
-    """A dictionary mapping output coordinate names to the retrieval rules and
-    preprocessing actions (e.g., DataConverters) that should be applied to each retrieved
-    coordinate variable."""
-
-    data_vars: Dict[str, Dict[Pattern, RetrievedVariable]]  # type: ignore
-    """A dictionary mapping output data variable names to the retrieval rules and
-    preprocessing actions (e.g., DataConverters) that should be applied to each
-    retrieved data variable."""
-
     def retrieve(
         self, input_keys: List[str], dataset_config: DatasetConfig, **kwargs: Any
     ) -> xr.Dataset:
@@ -90,11 +77,9 @@ class DefaultRetriever(Retriever):
         dataset_mapping: Dict[str, xr.Dataset] = {}
         for key, dataset in raw_mapping.items():
             input_config = InputKeyRetrieverConfig(key, self)
-            dataset = self._rename_variables(dataset, input_config)
-            dataset = self._reindex_dataset_coords(
-                dataset, dataset_config, input_config
-            )
-            dataset = self._run_data_converters(dataset, dataset_config, input_config)
+            dataset = _rename_variables(dataset, input_config)
+            dataset = _reindex_dataset_coords(dataset, dataset_config, input_config)
+            dataset = _run_data_converters(dataset, dataset_config, input_config)
             dataset_mapping[key] = dataset
         output_dataset = self._merge_raw_mapping(dataset_mapping)
         return output_dataset
@@ -120,133 +105,133 @@ class DefaultRetriever(Retriever):
                     break
         return input_reader_mapping
 
-    def _rename_variables(
-        self,
-        dataset: xr.Dataset,
-        input_config: InputKeyRetrieverConfig,
-    ) -> xr.Dataset:
-        """-----------------------------------------------------------------------------
-        Renames variables in the retrieved dataset according to retrieval configurations.
-
-        Args:
-            raw_dataset (xr.Dataset): The raw dataset.
-
-        Returns:
-            xr.Dataset: The simplified raw dataset.
-
-        -----------------------------------------------------------------------------"""
-        to_rename: Dict[str, str] = {}  # {raw_name: output_name}
-        coords_to_rename = {
-            c.name: output_name for output_name, c in input_config.coords.items()
-        }
-        vars_to_rename = {
-            v.name: output_name for output_name, v in input_config.data_vars.items()
-        }
-        to_rename.update(coords_to_rename)
-        to_rename.update(vars_to_rename)
-
-        for raw_name, output_name in coords_to_rename.items():
-            if raw_name not in dataset:
-                to_rename.pop(raw_name)
-                logger.warning(
-                    "Coordinate variable '%s' could not be retrieved from input. Please"
-                    " ensure the retrieval configuration file for the '%s' coord has"
-                    " the 'name' property set to the exact name of the variable in the"
-                    " dataset returned by the input DataReader.",
-                    raw_name,
-                    output_name,
-                )
-        for raw_name, output_name in vars_to_rename.items():
-            if raw_name not in dataset:
-                to_rename.pop(raw_name)
-                logger.warning(
-                    "Data variable '%s' could not be retrieved from input. Please"
-                    " ensure the retrieval configuration file for the '%s' data"
-                    " variable has the 'name' property set to the exact name of the"
-                    " variable in the dataset returned by the input DataReader.",
-                    raw_name,
-                    output_name,
-                )
-        return dataset.rename(to_rename)
-
-    def _run_data_converters(
-        self,
-        dataset: xr.Dataset,
-        dataset_config: DatasetConfig,
-        input_config: InputKeyRetrieverConfig,
-    ) -> xr.Dataset:
-        """------------------------------------------------------------------------------------
-        Runs the declared DataConverters on the dataset's coords and data_vars.
-
-        Returns the dataset after all converters have been run.
-
-        Args:
-            dataset (xr.Dataset): The dataset to convert.
-            dataset_config (DatasetConfig): The DatasetConfig
-
-        Returns:
-            xr.Dataset: The converted dataset.
-
-        ------------------------------------------------------------------------------------"""
-        for coord_name, coord_config in input_config.coords.items():
-            for converter in coord_config.data_converters:
-                dataset = converter.convert(dataset, dataset_config, coord_name)
-        for var_name, var_config in input_config.data_vars.items():
-            for converter in var_config.data_converters:
-                dataset = converter.convert(dataset, dataset_config, var_name)
-        return dataset
-
-    def _reindex_dataset_coords(
-        self,
-        dataset: xr.Dataset,
-        dataset_config: DatasetConfig,
-        input_config: InputKeyRetrieverConfig,
-    ) -> xr.Dataset:
-        """-----------------------------------------------------------------------------
-        Swaps dimensions and coordinates to match the structure of the DatasetConfig.
-
-        Ensures that the retriever coordinates are set as coordinates in the dataset,
-        promoting them to coordinates from data_vars as needed, and reindexes data_vars
-        so they are dimensioned by the appropriate coordinates.
-
-        This is useful in situations where the DataReader does not know which variables
-        to set as coordinates in its returned xr.Dataset, so it instead creates some
-        arbitrary index coordinate to dimension the data variables. This is very common
-        when reading from non-heirarchal formats such as csv.
-
-        Args:
-            dataset (xr.Dataset): The dataset to reindex.
-            dataset_config (DatasetConfig): The DatasetConfig.
-
-        Returns:
-            xr.Dataset: The reindexed dataset.
-
-        -----------------------------------------------------------------------------"""
-        for axis, coord_name in enumerate(input_config.coords):
-            expected_dim = dataset_config[coord_name].dims[0]
-            actual_dims = dataset[coord_name].dims
-            if (ndims := len(actual_dims)) > 1:
-                raise ValueError(
-                    f"Retrieved coordinate '{coord_name}' must have exactly one"
-                    f" dimension in the retrieved dataset, found {ndims} (dims="
-                    f"{actual_dims}). If '{coord_name}' is not actually a coordinate"
-                    " variable, please move it to the data_vars section in the"
-                    " retriever config file."
-                )
-            elif ndims == 0:
-                logger.warning(
-                    f"Retrieved coordinate '{coord_name}' has 0 attached dimensions in"
-                    " the retrieved dataset (expected ndims=1). Attempting to fix this"
-                    f" using xr.Dataset.expand_dims(dim='{coord_name}'), which may"
-                    " result in unexpected behavior. Please consider writing a"
-                    " DataReader to handle this coordinate correctly."
-                )
-                dataset = dataset.expand_dims(dim=coord_name, axis=axis)
-            dim = actual_dims[0] if ndims else coord_name
-            if dim != expected_dim:
-                dataset = dataset.swap_dims({dim: expected_dim})  # type: ignore
-
-        return dataset
-
     def _merge_raw_mapping(self, raw_mapping: Dict[str, xr.Dataset]) -> xr.Dataset:
         return xr.merge(list(raw_mapping.values()), **self.parameters.merge_kwargs)  # type: ignore
+
+
+def _rename_variables(
+    dataset: xr.Dataset,
+    input_config: InputKeyRetrieverConfig,
+) -> xr.Dataset:
+    """-----------------------------------------------------------------------------
+    Renames variables in the retrieved dataset according to retrieval configurations.
+
+    Args:
+        raw_dataset (xr.Dataset): The raw dataset.
+
+    Returns:
+        xr.Dataset: The simplified raw dataset.
+
+    -----------------------------------------------------------------------------"""
+    to_rename: Dict[str, str] = {}  # {raw_name: output_name}
+    coords_to_rename = {
+        c.name: output_name for output_name, c in input_config.coords.items()
+    }
+    vars_to_rename = {
+        v.name: output_name for output_name, v in input_config.data_vars.items()
+    }
+    to_rename.update(coords_to_rename)
+    to_rename.update(vars_to_rename)
+
+    for raw_name, output_name in coords_to_rename.items():
+        if raw_name not in dataset:
+            to_rename.pop(raw_name)
+            logger.warning(
+                "Coordinate variable '%s' could not be retrieved from input. Please"
+                " ensure the retrieval configuration file for the '%s' coord has"
+                " the 'name' property set to the exact name of the variable in the"
+                " dataset returned by the input DataReader.",
+                raw_name,
+                output_name,
+            )
+    for raw_name, output_name in vars_to_rename.items():
+        if raw_name not in dataset:
+            to_rename.pop(raw_name)
+            logger.warning(
+                "Data variable '%s' could not be retrieved from input. Please"
+                " ensure the retrieval configuration file for the '%s' data"
+                " variable has the 'name' property set to the exact name of the"
+                " variable in the dataset returned by the input DataReader.",
+                raw_name,
+                output_name,
+            )
+    return dataset.rename(to_rename)
+
+
+def _run_data_converters(
+    dataset: xr.Dataset,
+    dataset_config: DatasetConfig,
+    input_config: InputKeyRetrieverConfig,
+) -> xr.Dataset:
+    """------------------------------------------------------------------------------------
+    Runs the declared DataConverters on the dataset's coords and data_vars.
+
+    Returns the dataset after all converters have been run.
+
+    Args:
+        dataset (xr.Dataset): The dataset to convert.
+        dataset_config (DatasetConfig): The DatasetConfig
+
+    Returns:
+        xr.Dataset: The converted dataset.
+
+    ------------------------------------------------------------------------------------"""
+    for coord_name, coord_config in input_config.coords.items():
+        for converter in coord_config.data_converters:
+            dataset = converter.convert(dataset, dataset_config, coord_name)
+    for var_name, var_config in input_config.data_vars.items():
+        for converter in var_config.data_converters:
+            dataset = converter.convert(dataset, dataset_config, var_name)
+    return dataset
+
+
+def _reindex_dataset_coords(
+    dataset: xr.Dataset,
+    dataset_config: DatasetConfig,
+    input_config: InputKeyRetrieverConfig,
+) -> xr.Dataset:
+    """-----------------------------------------------------------------------------
+    Swaps dimensions and coordinates to match the structure of the DatasetConfig.
+
+    Ensures that the retriever coordinates are set as coordinates in the dataset,
+    promoting them to coordinates from data_vars as needed, and reindexes data_vars
+    so they are dimensioned by the appropriate coordinates.
+
+    This is useful in situations where the DataReader does not know which variables
+    to set as coordinates in its returned xr.Dataset, so it instead creates some
+    arbitrary index coordinate to dimension the data variables. This is very common
+    when reading from non-heirarchal formats such as csv.
+
+    Args:
+        dataset (xr.Dataset): The dataset to reindex.
+        dataset_config (DatasetConfig): The DatasetConfig.
+
+    Returns:
+        xr.Dataset: The reindexed dataset.
+
+    -----------------------------------------------------------------------------"""
+    for axis, coord_name in enumerate(input_config.coords):
+        expected_dim = dataset_config[coord_name].dims[0]
+        actual_dims = dataset[coord_name].dims
+        if (ndims := len(actual_dims)) > 1:
+            raise ValueError(
+                f"Retrieved coordinate '{coord_name}' must have exactly one"
+                f" dimension in the retrieved dataset, found {ndims} (dims="
+                f"{actual_dims}). If '{coord_name}' is not actually a coordinate"
+                " variable, please move it to the data_vars section in the"
+                " retriever config file."
+            )
+        elif ndims == 0:
+            logger.warning(
+                f"Retrieved coordinate '{coord_name}' has 0 attached dimensions in"
+                " the retrieved dataset (expected ndims=1). Attempting to fix this"
+                f" using xr.Dataset.expand_dims(dim='{coord_name}'), which may"
+                " result in unexpected behavior. Please consider writing a"
+                " DataReader to handle this coordinate correctly."
+            )
+            dataset = dataset.expand_dims(dim=coord_name, axis=axis)
+        dim = actual_dims[0] if ndims else coord_name
+        if dim != expected_dim:
+            dataset = dataset.swap_dims({dim: expected_dim})  # type: ignore
+
+    return dataset
